Delete tabs

+++ es.client/src/main/java/elastest/loganalyzer/es/client/rest/ExecutionRest.java
@@ -3,10 +3,11 @@ package elastest.loganalyzer.es.client.rest;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import elastest.loganalyzer.es.client.model.Execution;
@@ -14,33 +15,43 @@ import elastest.loganalyzer.es.client.service.ExecutionService;
 import elastest.loganalyzer.es.client.service.LogService;
 
 @RestController
-@RequestMapping("/executions")
+@RequestMapping("/api/executions")
 public class ExecutionRest {
 
-	private final ExecutionService executionService;
-	private final LogService logService;
-
 	@Autowired
-	public ExecutionRest(ExecutionService executionService, LogService logService) {
-		this.executionService = executionService;
-		this.logService = logService;
+	private ExecutionService executionService;
+	@Autowired
+	private LogService logService;
+
+	@RequestMapping(value = "/id/{id}", method = RequestMethod.GET)
+	public ResponseEntity<Execution> getById(@PathVariable int id) {
+		Execution execution = executionService.findOne(id);
+		if (execution != null) {
+			return new ResponseEntity<>(execution, HttpStatus.OK);	
+		} else {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
 	}
 
 	@RequestMapping(value = "/project/{project}", method = RequestMethod.GET)
-	public List<Execution> getByProject(@PathVariable String project,
-			@RequestParam(name = "tab", required = true) String tab) {
-		return executionService.findByProjectAndTabOrderById(project, tab);
-	}
-
-	@RequestMapping(value = "/test/{test}", method = RequestMethod.GET)
-	public Execution getByTest(@PathVariable String test) {
-		return executionService.findByTestId(test);
+	public ResponseEntity<List<Execution>> getByProject(@PathVariable String project) {
+		List<Execution> executions = executionService.findByProjectOrderById(project);
+		if (executions.size() > 0) {
+			return new ResponseEntity<>(executions, HttpStatus.OK);	
+		} else {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
 	}
 
 	@RequestMapping(value = "/id/{id}", method = RequestMethod.DELETE)
-	public void deleteById(@PathVariable int id) {
+	public ResponseEntity<Execution> deleteById(@PathVariable int id) {
 		Execution execution = executionService.findOne(id);
-		logService.deleteIterable(logService.findByTestOrderByIdAsc(execution.getTest()));
-		executionService.delete(execution);
+		if (execution != null) {
+			logService.deleteIterable(logService.findByTestOrderByIdAsc(String.format("%02d", execution.getId())));
+			executionService.delete(execution);
+			return new ResponseEntity<>(execution, HttpStatus.OK);
+		} else {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
 	}
 }
