+++ es.client/src/main/java/elastest/loganalyzer/es/client/rest/ProjectRest.java
@@ -4,60 +4,72 @@ import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
+import elastest.loganalyzer.es.client.model.Log;
 import elastest.loganalyzer.es.client.model.Project;
+import elastest.loganalyzer.es.client.service.ExecutionService;
 import elastest.loganalyzer.es.client.service.LogService;
 import elastest.loganalyzer.es.client.service.ProjectService;
-import elastest.loganalyzer.es.client.service.TabService;
 
 @RestController
-@RequestMapping("/projects")
+@RequestMapping("/api/projects")
 public class ProjectRest {
 
-	private final LogService logService;
-	private final ProjectService projectService;
-	private final TabService tabService;
-
 	@Autowired
-	public ProjectRest(LogService logService, ProjectService projectService, TabService tabService) {
-		this.logService = logService;
-		this.projectService = projectService;
-		this.tabService = tabService;
-	}
+	private ExecutionService executionService;
+	@Autowired
+	private LogService logService;
+	@Autowired
+	private ProjectService projectService;
 
-	@RequestMapping(value = "/all", method = RequestMethod.GET)
-	public List<Project> getAll() {
+	@RequestMapping(value = "", method = RequestMethod.GET)
+	public ResponseEntity<List<Project>> getAll() {
 		Iterable<Project> projects = projectService.findAll();
 		List<Project> result = new ArrayList<Project>();
 		for (Project project : projects) {
 			result.add(project);
 		}
-		return result;
+		if (result.size() > 0) {
+			return new ResponseEntity<>(result, HttpStatus.OK);	
+		} else {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
 	}
 
 	@RequestMapping(value = "/name/{name}", method = RequestMethod.GET)
-	public Project getProject(@PathVariable String name) {
-		return projectService.findByName(name);
+	public ResponseEntity<Project> getByName(@PathVariable String name) {
+		Project project = projectService.findByName(name);
+		if (project != null) {
+			return new ResponseEntity<>(project, HttpStatus.OK);	
+		} else {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
 	}
 
 	@RequestMapping(value = "", method = RequestMethod.POST, consumes = MediaType.APPLICATION_JSON_VALUE)
-	public int addLocation(@RequestBody Project project) {
+	public ResponseEntity<Integer> post(@RequestBody Project project) {
 		project.setRecently_deleted(-1);
-		return projectService.save(project);
+		return new ResponseEntity<>(projectService.save(project), HttpStatus.CREATED);
 	}
 
-	@RequestMapping(value = "/remove/id/{id}", method = RequestMethod.DELETE)
-	public Project delete(@PathVariable int id) {
+	@RequestMapping(value = "/id/{id}", method = RequestMethod.DELETE)
+	public ResponseEntity<Project> deleteById(@PathVariable int id) {
 		Project deleted = projectService.findOne(id);
-		logService.deleteIterable(logService.findByProject(deleted.getName()));
-		tabService.deleteIterable(tabService.findByProject(deleted.getName()));
-		projectService.delete(deleted);
-		return deleted;
+		if (deleted != null) {
+			logService.deleteIterable(logService.findByProject(deleted.getName()));
+			executionService.deleteIterable(executionService.findByProject(deleted.getName()));
+			projectService.delete(deleted);
+			return new ResponseEntity<>(deleted, HttpStatus.OK);
+		} else {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
 	}
 }
